requirements.txt
requests
psycopg2-binary
apache-airflow-providers-jdbc
mysql-connector-python
#openpyxl
boto3
#re
datetime
pandas